moved test for start and finish into notification subscriber tests
